move FormatError to errStack.
